Updated the DataSourceInterface
I've removed the getMood(), getPhoto() and getNotes() methods previously
moved in the EntryInterface class (see 7f369dc7ed3eafc3a27c99a18d2eda66191b7724)
I've also added the new getAvailableMoods() method that returns the list of
all Moods available for an Entry.
And, finally, I've replaced getEntries() with the slightly different
getDays() method that returns, instead of the list of all the avaible
entries, the list of days that have an Entry in the diary.