EntryAdapter displays the mood&note together.
Created the EntryAdapter for displaying the list of notes with the mood
icon. I must select an icon also when the user choose no mood, i decide to
put the ic_launcher icon. I also deleted the mood Icon from the
ListActivity's .xml and I created a button in WriteNote that links the
notes with the assignement of the mood.