First version of the rewrite of Photo.java
The old version did recreate a thumb everytime a Photo object was istantiated.
This new version only calculates the thumb_thumb path from the photo_path.
To get this one to work I need to rewrite a lot of things in DB.java and
PhotoActivity.java, wait for future commits.